Effortless attempts to fix Day 7

+++ Advent2015/Day/Day_07.h
@@ -21,8 +21,7 @@ namespace Day_07 {
 		std::string raw;
 		std::string id;
 		std::string operandName1, operandName2;
-		//unsigned short operand1, operand2, output;
-		int operand1, operand2, output;
+		unsigned short operand1, operand2, output;
 		unsigned char missingOperands;
 
 		inline IGate(const IGate & other) {};
